Refactor S3 buffer
