+++ ydb/core/tx/datashard/export_s3_buffer.h
@@ -5,14 +5,40 @@
 #include "export_iface.h"
 #include "export_scan.h"
 
+#include <util/generic/maybe.h>
+
 namespace NKikimr {
 namespace NDataShard {
 
-NExportScan::IBuffer* CreateS3ExportBufferRaw(
-    const IExport::TTableColumns& columns, ui64 maxRows, ui64 maxBytes, bool enableChecksums);
+struct TS3ExportBufferSettings {
+    struct TChecksumSettings {
+        enum class EChecksumType {
+            Sha256,
+        };
 
-NExportScan::IBuffer* CreateS3ExportBufferZstd(int compressionLevel,
-    const IExport::TTableColumns& columns, ui64 maxRows, ui64 maxBytes, ui64 minBytes, bool enableChecksums);
+        EChecksumType ChecksumType = EChecksumType::Sha256;
+    };
+
+    struct TCompressionSettings {
+        enum class ECompressionAlg {
+            Zstd,
+        };
+
+        ECompressionAlg Alg = ECompressionAlg::Zstd;
+        int CompressionLevel = -1;
+    };
+
+    IExport::TTableColumns Columns;
+    ui64 MaxRows = 0;
+    ui64 MinBytes = 0;
+    ui64 MaxBytes = 0;
+
+    // Data processing
+    TMaybe<TChecksumSettings> ChecksumSettings;
+    TMaybe<TCompressionSettings> CompressionSettings;
+};
+
+NExportScan::IBuffer* CreateS3ExportBuffer(TS3ExportBufferSettings&& settings);
 
 } // NDataShard
 } // NKikimr
