+++ ydb/core/tx/datashard/ya.make
@@ -293,8 +293,7 @@ IF (OS_WINDOWS)
     )
 ELSE()
     SRCS(
-        export_s3_buffer_raw.cpp
-        export_s3_buffer_zstd.cpp
+        export_s3_buffer.cpp
         export_s3_uploader.cpp
         extstorage_usage_config.cpp
         import_s3.cpp
